警告レベルを上昇 それに伴ってエラーを修正

+++ Engine/base/AudioManager/AudioManager.cpp
@@ -156,11 +156,11 @@ void AudioManager::StopAllSounds()
 	 MFCreateMediaType(&pMFMediaType);
 	 pMFMediaType->SetGUID(MF_MT_MAJOR_TYPE, MFMediaType_Audio);
 	 pMFMediaType->SetGUID(MF_MT_SUBTYPE, MFAudioFormat_PCM);
-	 pMFSourceReader->SetCurrentMediaType(MF_SOURCE_READER_FIRST_AUDIO_STREAM, nullptr, pMFMediaType);
+	 pMFSourceReader->SetCurrentMediaType((DWORD)MF_SOURCE_READER_FIRST_AUDIO_STREAM, nullptr, pMFMediaType);
 
 	 pMFMediaType->Release();
 	 pMFMediaType = nullptr;
-	 pMFSourceReader->GetCurrentMediaType(MF_SOURCE_READER_FIRST_AUDIO_STREAM, &pMFMediaType);
+	 pMFSourceReader->GetCurrentMediaType((DWORD)MF_SOURCE_READER_FIRST_AUDIO_STREAM, &pMFMediaType);
 
 	 WAVEFORMATEX *waveFormat{nullptr};
 	 MFCreateWaveFormatExFromMFMediaType(pMFMediaType, &waveFormat, nullptr);
@@ -171,7 +171,7 @@ void AudioManager::StopAllSounds()
 	 {
 		 IMFSample* pMFSample{ nullptr };
 		 DWORD dwStreamFlags{ 0 };
-		 pMFSourceReader->ReadSample(MF_SOURCE_READER_FIRST_AUDIO_STREAM, 0, nullptr, &dwStreamFlags, nullptr, &pMFSample);
+		 pMFSourceReader->ReadSample((DWORD)MF_SOURCE_READER_FIRST_AUDIO_STREAM, 0, nullptr, &dwStreamFlags, nullptr, &pMFSample);
 
 		 if (dwStreamFlags & MF_SOURCE_READERF_ENDOFSTREAM)
 		 {
